requirements.txt
urwid==1.3.1
paho-mqtt==1.2.1
